move the dependencies around
